add ( show_tour ) : add TourSeven and TourEight
